Renamed script_languages_container_ci_setup to exasol_script_languages_container_ci_setup
