+++ exasol_script_languages_container_ci_setup/cli/commands/deploy_ci_build.py
@@ -1,9 +1,9 @@
 import click
 
-from script_languages_container_ci_setup.cli.cli import cli
-from script_languages_container_ci_setup.cli.common import add_options
-from script_languages_container_ci_setup.lib.run_deploy_ci_build import run_deploy_ci_build
-from script_languages_container_ci_setup.cli.options.aws_options import aws_options
+from exasol_script_languages_container_ci_setup.cli.cli import cli
+from exasol_script_languages_container_ci_setup.cli.common import add_options
+from exasol_script_languages_container_ci_setup.lib.run_deploy_ci_build import run_deploy_ci_build
+from exasol_script_languages_container_ci_setup.cli.options.aws_options import aws_options
 
 
 CODE_BUILD_STACK_NAME = "CIBuild"
