+++ exasol_script_languages_container_ci_setup/cli/options/aws_options.py
