+++ exasol_script_languages_container_ci_setup/lib/run_deploy_ci_build.py
@@ -1,6 +1,6 @@
-from script_languages_container_ci_setup.lib.render_template import render_template
-from script_languages_container_ci_setup.lib.secret_name_reader import read_dockerhub_secret_name
-from script_languages_container_ci_setup.lib.upload_cloudformation import upload_cloudformation_stack
+from exasol_script_languages_container_ci_setup.lib.render_template import render_template
+from exasol_script_languages_container_ci_setup.lib.secret_name_reader import read_dockerhub_secret_name
+from exasol_script_languages_container_ci_setup.lib.upload_cloudformation import upload_cloudformation_stack
 
 
 def run_deploy_ci_build(aws_profile: str, project: str, github_url: str, stack_name: str):
