+++ exasol_script_languages_container_ci_setup/lib/run_generate_buildspec.py
@@ -4,7 +4,7 @@ from typing import Tuple
 import pkg_resources
 
 
-from script_languages_container_ci_setup.lib.render_template import render_template
+from exasol_script_languages_container_ci_setup.lib.render_template import render_template
 
 
 class Flavor(object):
