+++ exasol_script_languages_container_ci_setup/lib/secret_name_reader.py
