+++ exasol_script_languages_container_ci_setup/lib/upload_cloudformation.py
