+++ exasol_script_languages_container_ci_setup/templates/build_buildspec.yaml
