+++ exasol_script_languages_container_ci_setup/templates/buildspec_batch_entry.yaml
