+++ exasol_script_languages_container_ci_setup/templates/slc_source_credential.yaml
