+++ pyproject.toml
@@ -1,5 +1,5 @@
 [tool.poetry]
-name = "script-languages-container-ci-setup"
+name = "exasol-script-languages-container-ci-setup"
 version = "0.1.0"
 description = "Manages AWS cloud CI build infrastructure."
 
@@ -14,7 +14,7 @@ python = ">=3.8.0,<4.0"
 click = "^8.0.3"
 jinja2 = ">=3.0.0"
 exasol_error_reporting_python = { git = "https://github.com/exasol/error-reporting-python.git", branch = "main" }
-script-languages-container-ci = {git = "https://github.com/exasol/script-languages-container-ci.git", branch="feature/1_add_initial_implementation_of_slc_ci_project"}
+exasol_script_languages_container_tool = { url = "https://github.com/exasol/script-languages-container-tool/releases/download/0.10.0/exasol_script_languages_container_tool-0.10.0-py3-none-any.whl" }
 
 [build-system]
 requires = ["poetry_core>=1.0.0"]
@@ -23,3 +23,13 @@ build-backend = "poetry.core.masonry.api"
 
 [tool.poetry.dev-dependencies]
 toml = ">=0.10.2"
+pytest = "^7.1.1"
+pytest-mock = "^3.7.0"
+
+
+[tool.pytest.ini_options]
+minversion = "6.0"
+
+testpaths = [
+    "test"
+]
