--- script_languages_container_ci_setup/lib/render_template.py
@@ -1,7 +1,0 @@
-import jinja2
-
-
-def render_template(template: str, **kwargs):
-    env = jinja2.Environment(loader=jinja2.PackageLoader("script_languages_container_ci_setup"), autoescape=jinja2.select_autoescape())
-    t = env.get_template(template)
-    return t.render(**kwargs)
