+++ setup.py
@@ -2,25 +2,25 @@
 from setuptools import setup
 
 packages = \
-['script_languages_container_ci_setup',
- 'script_languages_container_ci_setup.cli',
- 'script_languages_container_ci_setup.cli.commands',
- 'script_languages_container_ci_setup.cli.options',
- 'script_languages_container_ci_setup.lib']
+['exasol_script_languages_container_ci_setup',
+ 'exasol_script_languages_container_ci_setup.cli',
+ 'exasol_script_languages_container_ci_setup.cli.commands',
+ 'exasol_script_languages_container_ci_setup.cli.options',
+ 'exasol_script_languages_container_ci_setup.lib']
 
 package_data = \
-{'': ['*'], 'script_languages_container_ci_setup': ['templates/*']}
+{'': ['*'], 'exasol_script_languages_container_ci_setup': ['templates/*']}
 
 install_requires = \
 ['click>=8.0.3,<9.0.0',
  'exasol_error_reporting_python @ '
  'git+https://github.com/exasol/error-reporting-python.git@main',
- 'jinja2>=3.0.0',
- 'script-languages-container-ci @ '
- 'git+https://github.com/exasol/script-languages-container-ci.git@feature/1_add_initial_implementation_of_slc_ci_project']
+ 'exasol_script_languages_container_tool @ '
+ 'https://github.com/exasol/script-languages-container-tool/releases/download/0.10.0/exasol_script_languages_container_tool-0.10.0-py3-none-any.whl',
+ 'jinja2>=3.0.0']
 
 setup_kwargs = {
-    'name': 'script-languages-container-ci-setup',
+    'name': 'exasol-script-languages-container-ci-setup',
     'version': '0.1.0',
     'description': 'Manages AWS cloud CI build infrastructure.',
     'long_description': None,
